refactor: moved SwitchCondition into the separate component (M2-6691)
