+++ src/modules/Builder/components/ConditionRow/Condition/Condition.styles.ts
@@ -1,7 +1,7 @@
 import { styled } from '@mui/material';
 
 import { StyledFlexTopCenter, variables, theme } from 'shared/styles';
-import { SelectController, InputController } from 'shared/components/FormComponents';
+import { SelectController } from 'shared/components/FormComponents';
 
 export const StyledCondition = styled(StyledFlexTopCenter)`
   gap: 0.4rem;
@@ -61,35 +61,6 @@ export const StyledSelectController = styled(SelectController)`
   }
 
   .MuiSvgIcon-root.Mui-disabled {
-    display: none;
-  }
-
-  .MuiOutlinedInput-notchedOutline {
-    border-color: ${variables.palette.outline};
-  }
-`;
-
-export const StyledInputController = styled(InputController)`
-  width: 8rem;
-
-  .MuiInputBase-root {
-    border-radius: ${variables.borderRadius.md};
-    padding-right: ${theme.spacing(0.7)};
-
-    &.Mui-error {
-      background: ${variables.palette.error_container};
-    }
-  }
-
-  .MuiInputBase-input {
-    padding: ${theme.spacing(0.6, 1.2)};
-  }
-
-  .MuiBox-root {
-    margin-left: 0;
-  }
-
-  .MuiFormHelperText-root {
     display: none;
   }
