+++ src/modules/Builder/components/ConditionRow/Condition/Condition.types.ts
@@ -30,9 +30,3 @@ export type ConditionProps = {
   type: ConditionRowType;
   'data-testid'?: string;
 };
-
-export type SwitchConditionProps = {
-  selectedItem?: ConditionItem;
-  itemType?: ConditionItemType;
-  dataTestid: ConditionProps['data-testid'];
-} & Pick<ConditionProps, 'numberValueName' | 'minValueName' | 'maxValueName' | 'state'>;
