+++ src/modules/Builder/components/ConditionRow/Condition/SwitchCondition/SwitchCondition.types.ts
@@ -1,0 +1,8 @@
+import { ConditionItemType } from '../Condition.const';
+import { ConditionItem, ConditionProps } from '../Condition.types';
+
+export type SwitchConditionProps = {
+  selectedItem?: ConditionItem;
+  itemType?: ConditionItemType;
+  dataTestid: ConditionProps['data-testid'];
+} & Pick<ConditionProps, 'numberValueName' | 'minValueName' | 'maxValueName' | 'state'>;
