+++ src/modules/Builder/components/ConditionRow/Condition/SwitchCondition/index.ts
@@ -1,0 +1,1 @@
+export * from './SwitchCondition';
